"UI split into classes: Display, TileButton and UI to unspaghettise code. UI and Display are initialize in GameMaster For now right-click on a tile will spawn a random bug and left-click will select bug's army"
